Replace int with double as a parser priority

+++ panda-framework/src/main/java/org/panda_lang/panda/framework/design/interpreter/parser/pipeline/ParserRepresentation.java
@@ -33,7 +33,7 @@ public interface ParserRepresentation {
     /**
      * @return priority
      */
-    int getPriority();
+    double getPriority();
 
     /**
      * @return associated handler
